requirements.txt
BeautifulSoup4==4.3.2
Flask==0.10.1
Flask-Admin==1.0.8
Flask-BabelEx==0.9.2
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.3.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Security==1.7.4
Flask-WTF==0.11
Markdown==2.5.2
qiniu==7.0.2